add 1 model serving
